+++ packages/lumx-react/src/components/skeleton/index.ts
@@ -1,0 +1,3 @@
+export * from './SkeletonCircle';
+export * from './SkeletonRectangle';
+export * from './SkeletonTypography';