+++ packages/site-demo/content/product/components/skeleton/react/circle.tsx
@@ -1,0 +1,12 @@
+import { Size, SkeletonCircle } from '@lumx/react';
+import React from 'react';
+
+export const App = ({ theme }: any) => {
+    return (
+        <>
+            <SkeletonCircle theme={theme} size={Size.m} />
+            <SkeletonCircle theme={theme} size={Size.l} />
+            <SkeletonCircle theme={theme} size={Size.xl} />
+        </>
+    );
+};