+++ CHANGELOG.md
@@ -61,6 +61,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   _[BREAKING]_ Removed `Tabs` component replaced by the `TabProvider` and `TabList` components.
 -   _[BREAKING]_ Removed `thumbnails.url` prop on the `Mosaic` (use `image` instead).
 
+## [0.28.2][] - 2020-12-11
+
+### Added
+
+-   Added `headerActions` prop to `CommentBlock` component to add actions to the header.
+-   Added `SkeletonRectangle`, `SkeletonCircle` and `SkeletonTypography` components.
+
+### Fixed
+
+-   Fixed `text` prop of the `CommentBlock` component to accept `ReactNode` type.
+
 ## [0.28.1][] - 2020-12-03
 
 ### Fixed
@@ -863,3 +874,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.28.0]: https://github.com/lumapps/design-system/tree/v0.28.0
 [unreleased]: https://github.com/lumapps/design-system/compare/v0.28.1...HEAD
 [0.28.1]: https://github.com/lumapps/design-system/tree/v0.28.1
+[unreleased]: https://github.com/lumapps/design-system/compare/v0.28.2...HEAD
+[0.28.2]: https://github.com/lumapps/design-system/tree/v0.28.2